+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/TruffleFile.java
@@ -95,6 +95,11 @@ import java.util.Set;
 import org.graalvm.polyglot.io.FileSystem;
 
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
+import com.oracle.truffle.api.impl.TruffleLocator;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.ServiceLoader;
+import java.util.function.Supplier;
 
 /**
  * An abstract representation of a file used by Truffle languages.
@@ -104,20 +109,21 @@ import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 public final class TruffleFile {
     private static final int MAX_BUFFER_SIZE = Integer.MAX_VALUE - 8;
     private static final int BUFFER_SIZE = 8192;
+    private static final boolean LEGACY_FILETYPEDETECTORS_DISABLED = Boolean.getBoolean("graalvm.legacy.filetypedetectors.disabled");
 
-    private final FileSystem fileSystem;
+    private final FileSystemContext fileSystemContext;
     private final Path path;
     private final Path normalizedPath;
 
-    TruffleFile(final FileSystem fileSystem, final Path path) {
-        this(fileSystem, path, path.normalize());
+    TruffleFile(final FileSystemContext fileSystemContext, final Path path) {
+        this(fileSystemContext, path, path.normalize());
     }
 
-    TruffleFile(final FileSystem fileSystem, final Path path, final Path normalizedPath) {
-        Objects.requireNonNull(fileSystem, "FileSystem must not be null.");
+    TruffleFile(final FileSystemContext fileSystemContext, final Path path, final Path normalizedPath) {
+        Objects.requireNonNull(fileSystemContext, "FileSystemContext must not be null.");
         Objects.requireNonNull(path, "Path must not be null.");
         Objects.requireNonNull(normalizedPath, "NormalizedPath must not be null.");
-        this.fileSystem = fileSystem;
+        this.fileSystemContext = fileSystemContext;
         this.path = path;
         this.normalizedPath = normalizedPath;
     }
@@ -346,7 +352,7 @@ public final class TruffleFile {
             return toUri();
         }
         try {
-            String strPath = "/".equals(fileSystem.getSeparator()) ? path.toString() : path.toString().replace(path.getFileSystem().getSeparator(), "/");
+            String strPath = "/".equals(fileSystemContext.fileSystem.getSeparator()) ? path.toString() : path.toString().replace(path.getFileSystem().getSeparator(), "/");
             return new URI(null, null, strPath, null);
         } catch (Throwable t) {
             throw wrapHostException(t);
@@ -369,7 +375,7 @@ public final class TruffleFile {
         }
         try {
             Path[] absolutePaths = toAbsolutePathImpl();
-            return new TruffleFile(fileSystem, absolutePaths[0], absolutePaths[1]);
+            return new TruffleFile(fileSystemContext, absolutePaths[0], absolutePaths[1]);
         } catch (SecurityException se) {
             throw se;
         } catch (Throwable t) {
@@ -389,7 +395,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public TruffleFile getCanonicalFile(LinkOption... options) throws IOException {
         try {
-            return new TruffleFile(fileSystem, fileSystem.toRealPath(path, options));
+            return new TruffleFile(fileSystemContext, fileSystemContext.fileSystem.toRealPath(path, options));
         } catch (IOException | SecurityException e) {
             throw e;
         } catch (Throwable t) {
@@ -407,7 +413,7 @@ public final class TruffleFile {
     public TruffleFile getParent() {
         try {
             final Path parent = path.getParent();
-            return parent == null ? null : new TruffleFile(fileSystem, parent);
+            return parent == null ? null : new TruffleFile(fileSystemContext, parent);
         } catch (Throwable t) {
             throw wrapHostException(t);
         }
@@ -424,7 +430,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public TruffleFile resolve(String name) {
         try {
-            return new TruffleFile(fileSystem, path.resolve(name));
+            return new TruffleFile(fileSystemContext, path.resolve(name));
         } catch (InvalidPathException ip) {
             throw ip;
         } catch (Throwable t) {
@@ -443,7 +449,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public TruffleFile resolveSibling(String name) {
         try {
-            return new TruffleFile(fileSystem, path.resolveSibling(name));
+            return new TruffleFile(fileSystemContext, path.resolveSibling(name));
         } catch (InvalidPathException ip) {
             throw ip;
         } catch (Throwable t) {
@@ -503,7 +509,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public void setLastModifiedTime(FileTime time, LinkOption... options) throws IOException {
         try {
-            fileSystem.setAttribute(normalizedPath, "lastModifiedTime", time, options);
+            fileSystemContext.fileSystem.setAttribute(normalizedPath, "lastModifiedTime", time, options);
         } catch (IOException | SecurityException e) {
             throw e;
         } catch (Throwable t) {
@@ -543,7 +549,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public void setLastAccessTime(FileTime time, LinkOption... options) throws IOException {
         try {
-            fileSystem.setAttribute(normalizedPath, "lastAccessTime", time, options);
+            fileSystemContext.fileSystem.setAttribute(normalizedPath, "lastAccessTime", time, options);
         } catch (IOException | SecurityException e) {
             throw e;
         } catch (Throwable t) {
@@ -583,7 +589,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public void setCreationTime(FileTime time, LinkOption... options) throws IOException {
         try {
-            fileSystem.setAttribute(normalizedPath, "creationTime", time, options);
+            fileSystemContext.fileSystem.setAttribute(normalizedPath, "creationTime", time, options);
         } catch (IOException | SecurityException e) {
             throw e;
         } catch (Throwable t) {
@@ -606,10 +612,10 @@ public final class TruffleFile {
         try {
             final Collection<TruffleFile> result = new ArrayList<>();
             final boolean normalized = isNormalized();
-            try (DirectoryStream<Path> stream = fileSystem.newDirectoryStream(normalizedPath, AllFiles.INSTANCE)) {
+            try (DirectoryStream<Path> stream = fileSystemContext.fileSystem.newDirectoryStream(normalizedPath, AllFiles.INSTANCE)) {
                 for (Path p : stream) {
                     result.add(new TruffleFile(
-                                    fileSystem,
+                                    fileSystemContext,
                                     normalized ? p : path.resolve(p.getFileName()),
                                     normalized ? p : normalizedPath.resolve(p.getFileName())));
                 }
@@ -640,7 +646,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public SeekableByteChannel newByteChannel(Set<? extends OpenOption> options, FileAttribute<?>... attributes) throws IOException {
         try {
-            return ByteChannelDecorator.create(fileSystem.newByteChannel(normalizedPath, options, attributes));
+            return ByteChannelDecorator.create(fileSystemContext.fileSystem.newByteChannel(normalizedPath, options, attributes));
         } catch (IOException | UnsupportedOperationException | IllegalArgumentException | SecurityException e) {
             throw e;
         } catch (Throwable t) {
@@ -872,7 +878,7 @@ public final class TruffleFile {
             SecurityException notAllowed = null;
             Path absolutePath = normalizedPath;
             try {
-                absolutePath = fileSystem.toAbsolutePath(absolutePath);
+                absolutePath = fileSystemContext.fileSystem.toAbsolutePath(absolutePath);
             } catch (SecurityException se) {
                 notAllowed = se;
             }
@@ -909,7 +915,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public void delete() throws IOException {
         try {
-            fileSystem.delete(normalizedPath);
+            fileSystemContext.fileSystem.delete(normalizedPath);
         } catch (IOException | SecurityException e) {
             throw e;
         } catch (Throwable t) {
@@ -938,7 +944,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public void move(TruffleFile target, CopyOption... options) throws IOException {
         try {
-            fileSystem.move(normalizedPath, target.normalizedPath, options);
+            fileSystemContext.fileSystem.move(normalizedPath, target.normalizedPath, options);
         } catch (IOException | UnsupportedOperationException | SecurityException e) {
             throw e;
         } catch (Throwable t) {
@@ -983,7 +989,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public void setPosixPermissions(Set<? extends PosixFilePermission> permissions, LinkOption... linkOptions) throws IOException {
         try {
-            fileSystem.setAttribute(normalizedPath, "posix:permissions", permissions, linkOptions);
+            fileSystemContext.fileSystem.setAttribute(normalizedPath, "posix:permissions", permissions, linkOptions);
         } catch (IOException | SecurityException | UnsupportedOperationException e) {
             throw e;
         } catch (Throwable t) {
@@ -1011,7 +1017,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public int hashCode() {
         int res = 17;
-        res = res * 31 + fileSystem.hashCode();
+        res = res * 31 + fileSystemContext.hashCode();
         res = res * 31 + path.hashCode();
         return res;
     }
@@ -1031,7 +1037,7 @@ public final class TruffleFile {
             return false;
         }
         final TruffleFile otherFile = (TruffleFile) other;
-        return path.equals(otherFile.path) && fileSystem.equals(otherFile.fileSystem);
+        return path.equals(otherFile.path) && fileSystemContext.equals(otherFile.fileSystemContext);
     }
 
     /**
@@ -1045,7 +1051,7 @@ public final class TruffleFile {
         if (isNormalized()) {
             return this;
         }
-        return new TruffleFile(fileSystem, normalizedPath, normalizedPath);
+        return new TruffleFile(fileSystemContext, normalizedPath, normalizedPath);
     }
 
     /**
@@ -1069,7 +1075,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public TruffleFile relativize(TruffleFile other) {
         try {
-            return new TruffleFile(fileSystem, path.relativize(other.path));
+            return new TruffleFile(fileSystemContext, path.relativize(other.path));
         } catch (IllegalArgumentException e) {
             throw e;
         } catch (Throwable t) {
@@ -1168,7 +1174,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public void createLink(TruffleFile target) throws IOException {
         try {
-            fileSystem.createLink(normalizedPath, target.normalizedPath);
+            fileSystemContext.fileSystem.createLink(normalizedPath, target.normalizedPath);
         } catch (IOException | SecurityException | UnsupportedOperationException e) {
             throw e;
         } catch (Throwable t) {
@@ -1192,7 +1198,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public void createSymbolicLink(TruffleFile target, FileAttribute<?>... attrs) throws IOException {
         try {
-            fileSystem.createSymbolicLink(normalizedPath, target.path, attrs);
+            fileSystemContext.fileSystem.createSymbolicLink(normalizedPath, target.path, attrs);
         } catch (IOException | SecurityException | UnsupportedOperationException e) {
             throw e;
         } catch (Throwable t) {
@@ -1265,7 +1271,7 @@ public final class TruffleFile {
     @TruffleBoundary
     public DirectoryStream<TruffleFile> newDirectoryStream() throws IOException {
         try {
-            return new TruffleFileDirectoryStream(this, fileSystem.newDirectoryStream(normalizedPath, AllFiles.INSTANCE));
+            return new TruffleFileDirectoryStream(this, fileSystemContext.fileSystem.newDirectoryStream(normalizedPath, AllFiles.INSTANCE));
         } catch (IOException | SecurityException e) {
             throw e;
         } catch (Throwable t) {
@@ -1391,11 +1397,167 @@ public final class TruffleFile {
     @TruffleBoundary
     public void copy(TruffleFile target, CopyOption... options) throws IOException {
         try {
-            fileSystem.copy(normalizedPath, target.normalizedPath, options);
+            fileSystemContext.fileSystem.copy(normalizedPath, target.normalizedPath, options);
         } catch (IOException | UnsupportedOperationException | SecurityException e) {
             throw e;
         } catch (Throwable t) {
             throw wrapHostException(t);
+        }
+    }
+
+    /**
+     * Returns the {@link TruffleFile file} MIME type.
+     *
+     * @return the MIME type or {@code null} if the MIME type is not recognized
+     * @throws IOException in case of IO error
+     * @throws SecurityException if the {@link FileSystem} denied the operation
+     * @since 1.0
+     */
+    @TruffleBoundary
+    public String getMimeType() throws IOException {
+        return getMimeType(null);
+    }
+
+    @TruffleBoundary
+    String getMimeType(Set<String> validMimeTypes) throws IOException {
+        try {
+            String result = fileSystemContext.fileSystem.getMimeType(normalizedPath);
+            if (result != null && (validMimeTypes == null || validMimeTypes.contains(result))) {
+                return result;
+            }
+            for (FileTypeDetector detector : fileSystemContext.getFileTypeDetectors()) {
+                result = detector.findMimeType(this);
+                if (result != null && (validMimeTypes == null || validMimeTypes.contains(result))) {
+                    return result;
+                }
+            }
+            return getMimeTypeLegacy(path.toString(), validMimeTypes);
+        } catch (IOException | SecurityException e) {
+            throw e;
+        } catch (Throwable t) {
+            throw wrapHostException(t);
+        }
+    }
+
+    private static String getMimeTypeLegacy(String path, Set<String> validMimeTypes) throws IOException {
+        if (LEGACY_FILETYPEDETECTORS_DISABLED) {
+            return null;
+        } else {
+            try {
+                Path legacyPath = Paths.get(path);
+                if (!TruffleOptions.AOT) {
+                    Collection<ClassLoader> loaders = TruffleLocator.loaders();
+                    for (ClassLoader l : loaders) {
+                        for (java.nio.file.spi.FileTypeDetector detector : ServiceLoader.load(java.nio.file.spi.FileTypeDetector.class, l)) {
+                            String mimeType = detector.probeContentType(legacyPath);
+                            if (mimeType != null && (validMimeTypes == null || validMimeTypes.contains(mimeType))) {
+                                return mimeType;
+                            }
+                        }
+                    }
+                }
+                String contentType = Files.probeContentType(legacyPath);
+                if (contentType != null && (validMimeTypes == null || validMimeTypes.contains(contentType))) {
+                    return contentType;
+                }
+                return null;
+            } catch (InvalidPathException e) {
+                return null;
+            }
+        }
+    }
+
+    /**
+     * Returns the {@link TruffleFile file} encoding. For a file containing an encoding information
+     * returns the encoding.
+     *
+     * @return the file encoding or {@code null} if the file does not provide encoding
+     * @throws IOException in case of IO error
+     * @throws SecurityException if the {@link FileSystem} denied the operation
+     * @since 1.0
+     */
+    @TruffleBoundary
+    public Charset getEncoding() throws IOException {
+        try {
+            Charset result = fileSystemContext.fileSystem.getEncoding(normalizedPath);
+            if (result != null) {
+                return result;
+            }
+            for (FileTypeDetector detector : fileSystemContext.getFileTypeDetectors()) {
+                result = detector.findEncoding(this);
+                if (result != null) {
+                    return result;
+                }
+            }
+            return null;
+        } catch (IOException | UnsupportedOperationException | SecurityException e) {
+            throw e;
+        } catch (Throwable t) {
+            throw wrapHostException(t);
+        }
+    }
+
+    /**
+     * A detector for finding {@link TruffleFile file}'s MIME type and encoding.
+     *
+     * <p>
+     * The means by which the detector determines the MIME is highly implementation specific. A
+     * simple implementation might detect the MIME type using {@link TruffleFile file} extension. In
+     * other cases, the content of {@link TruffleFile file} needs to be examined to guess its file
+     * type.
+     * <p>
+     * The implementations are registered using
+     * {@link TruffleLanguage.Registration#fileTypeDetectors() TruffleLanguage registration}.
+     *
+     * @see TruffleFile#getMimeType()
+     * @see TruffleLanguage.Registration#fileTypeDetectors()
+     * @since 1.0
+     */
+    public interface FileTypeDetector {
+        /**
+         * Finds a MIME type for given {@link TruffleFile}.
+         *
+         * @param file the {@link TruffleFile file} to find a MIME type for
+         * @return the MIME type or {@code null} if the MIME type is not recognized
+         * @throws IOException of an I/O error occurs
+         * @throws SecurityException if the implementation requires an access the file and the
+         *             {@link FileSystem} denies the operation
+         * @since 1.0
+         */
+        String findMimeType(TruffleFile file) throws IOException;
+
+        /**
+         * For a file containing an encoding information returns the encoding.
+         *
+         * @param file the {@link TruffleFile file} to find an encoding for
+         * @return the file encoding or {@code null} if the file does not provide encoding
+         * @throws IOException of an I/O error occurs
+         * @throws SecurityException if the {@link FileSystem} denies the file access
+         * @since 1.0
+         */
+        Charset findEncoding(TruffleFile file) throws IOException;
+    }
+
+    static final class FileSystemContext {
+        final FileSystem fileSystem;
+        private final Supplier<Iterable<? extends FileTypeDetector>> fileTypeDetectorsSupplier;
+        private volatile Iterable<? extends FileTypeDetector> fileTypeDetectors;
+
+        FileSystemContext(FileSystem fileSystem, Supplier<Iterable<? extends FileTypeDetector>> fileTypeDetectorsSupplier) {
+            Objects.requireNonNull(fileSystem, "FileSystem must be non null.");
+            Objects.requireNonNull(fileTypeDetectorsSupplier, "FileTypeDetectorsSupplier must be non null.");
+            this.fileSystem = fileSystem;
+            this.fileTypeDetectorsSupplier = fileTypeDetectorsSupplier;
+        }
+
+        Iterable<? extends FileTypeDetector> getFileTypeDetectors() {
+            Iterable<? extends FileTypeDetector> result = fileTypeDetectors;
+            if (result == null) {
+                result = fileTypeDetectorsSupplier.get();
+                assert result != null : "FileTypeDetectorsSupplier returned null.";
+                fileTypeDetectors = result;
+            }
+            return result;
         }
     }
 
@@ -1404,8 +1566,8 @@ public final class TruffleFile {
     }
 
     private Path[] toAbsolutePathImpl() {
-        Path absolute = fileSystem.toAbsolutePath(path);
-        Path normalizedAbsolute = fileSystem.toAbsolutePath(normalizedPath).normalize();
+        Path absolute = fileSystemContext.fileSystem.toAbsolutePath(path);
+        Path normalizedAbsolute = fileSystemContext.fileSystem.toAbsolutePath(normalizedPath).normalize();
         return new Path[]{absolute, normalizedAbsolute};
     }
 
@@ -1417,7 +1579,7 @@ public final class TruffleFile {
 
     private boolean checkAccess(Set<? extends AccessMode> modes, LinkOption... linkOptions) {
         try {
-            fileSystem.checkAccess(normalizedPath, modes, linkOptions);
+            fileSystemContext.fileSystem.checkAccess(normalizedPath, modes, linkOptions);
             return true;
         } catch (IOException ioe) {
             return false;
@@ -1434,14 +1596,14 @@ public final class TruffleFile {
     }
 
     private Object getAttributeImpl(final Path forPath, final String attribute, final LinkOption... options) throws IOException {
-        final Map<String, Object> map = fileSystem.readAttributes(forPath, attribute, options);
+        final Map<String, Object> map = fileSystemContext.fileSystem.readAttributes(forPath, attribute, options);
         final int index = attribute.indexOf(':');
         final String key = index < 0 ? attribute : attribute.substring(index + 1);
         return map.get(key);
     }
 
     private Path createDirectoryImpl(Path dir, FileAttribute<?>... attrs) throws IOException {
-        fileSystem.createDirectory(dir, attrs);
+        fileSystemContext.fileSystem.createDirectory(dir, attrs);
         return dir;
     }
 
@@ -1465,7 +1627,7 @@ public final class TruffleFile {
         final Set<AccessMode> mode = EnumSet.noneOf(AccessMode.class);
         for (Path p = forPath.getParent(); p != null; p = p.getParent()) {
             try {
-                fileSystem.checkAccess(p, mode);
+                fileSystemContext.fileSystem.checkAccess(p, mode);
                 return p;
             } catch (NoSuchFileException nsfe) {
                 // Still does not exist
@@ -1475,7 +1637,7 @@ public final class TruffleFile {
     }
 
     private <T extends Throwable> RuntimeException wrapHostException(T t) {
-        throw wrapHostException(t, fileSystem);
+        throw wrapHostException(t, fileSystemContext.fileSystem);
     }
 
     static <T extends Throwable> RuntimeException wrapHostException(T t, FileSystem fs) {
@@ -1616,7 +1778,7 @@ public final class TruffleFile {
                 try {
                     Path path = delegateIterator.next();
                     return new TruffleFile(
-                                    directory.fileSystem,
+                                    directory.fileSystemContext,
                                     normalized ? path : directory.path.resolve(path.getFileName()),
                                     normalized ? path : directory.normalizedPath.resolve(path.getFileName()));
                 } catch (DirectoryIteratorException e) {
@@ -1761,7 +1923,7 @@ public final class TruffleFile {
             private Event enter(TruffleFile file) {
                 BasicFileAttributes attrs;
                 try {
-                    attrs = new BasicFileAttributesImpl(file.fileSystem.readAttributes(file.normalizedPath, "*", linkOptions));
+                    attrs = new BasicFileAttributesImpl(file.fileSystemContext.fileSystem.readAttributes(file.normalizedPath, "*", linkOptions));
                 } catch (IOException ioe) {
                     return new Event(Event.Type.VISIT, file, ioe);
                 }